remove localhost ref, move columnData.js, add env for DB
