+++ lib/dbConnect.js
@@ -1,6 +1,7 @@
 import mongoose from 'mongoose';
 
 const MONGODB_URI = process.env.MONGODB_URI;
+const MONGODB_ENV = process.env.MONGODB_ENV;
 
 if (!MONGODB_URI) {
   throw new Error(
@@ -27,6 +28,7 @@ async function dbConnect() {
   if (!cached.promise) {
     const opts = {
       bufferCommands: false,
+      dbName: MONGODB_ENV,
     };
 
     cached.promise = mongoose.connect(MONGODB_URI, opts).then((mongoose) => {
